Update script so that conda builds CGNS
It was a struggle to help CGNS find the hdf5 and zlib libraries, but
this partly may have been due to my confusion with the conda build
process. Also, setting `-D_LARGEFILE64_SOURCE` was key.

Unfortuantely, 16 of the 19 tests fail. I'm going to pass on fixing
them now in the hopes that if I ignore this, it won't be a
problem. I'll add an issue, though.